refactor: moving according colors into utils file
